Attack Range Fix

+++ RPG Game/Assets/_Scripts/Player/WeaponScriptableObject.cs
@@ -16,7 +16,7 @@ public class WeaponScriptableObject : ScriptableObject
     [Header("Weapon")]
     public int weaponIndex;
     public int attackDamage;
-    public int attackRange;
+    public float attackRange;
 
     [Header("Basic Attack Ability")]
     public float basicAttackCoolDown;
